+++ src/any.rs
@@ -117,7 +117,7 @@ impl From<AnyOmeZarrAttributes> for AnyOmeFields {
 
 #[cfg(test)]
 mod tests {
-    use serde::{de::DeserializeOwned, Serialize};
+    use serde::{Serialize, de::DeserializeOwned};
     use validatrix::Valid;
 
     use super::*;