+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [Unreleased](https://github.com/zarrs/ome_zarr_metadata/compare/v0.3.0...HEAD)
+
+## [0.3.0](https://github.com/zarrs/ome_zarr_metadata/compare/v0.2.5...v0.3.0) - 2026-03-31
 
 ### Added
 
@@ -106,7 +108,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial public release
 
-[unreleased]: https://github.com/zarrs/ome_zarr_metadata/compare/v0.2.5...HEAD
 [0.2.5]: https://github.com/zarrs/ome_zarr_metadata/releases/tag/v0.2.5
 [0.2.4]: https://github.com/zarrs/ome_zarr_metadata/releases/tag/v0.2.4
 [0.2.3]: https://github.com/zarrs/ome_zarr_metadata/releases/tag/v0.2.3